# perf: very slow startup time caused by expo-module using kotlin reflection