# Why

fix https://github.com/expo/expo/issues/39368


I'm investigating slow app startup on Asus zenfone 3 max and I've noticed that expo modules are part of the reason of the slow startup.

Using react-native 79, expo 53.

For a ~4s startup time expo module is responsible for at least 1/4 to 1/3 of this time.
*This is when using release mode (not debug)*

<img width="1496" height="466" alt="Image" src="https://github.com/user-attachments/assets/3175ba54-f935-456f-9e6d-aed65b8a4473" />

Since I wasn't sure about the correctness of the profiling session I added logs in react-native and expo.

I also added log in this turboModuleManager initialize function 
https://github.com/facebook/react-native/blob/f9cecc5f00f0a6008d32ad2ff2e8d3cc654e9c58/packages/react-native/ReactAndroid/src/main/java/com/facebook/react/ReactPackageTurboModuleManagerDelegate.kt#L48

and another one in the expo ModuleRegistry register function

https://github.com/expo/expo/blob/ef0a0714612c11e10b8a92c9fb023066be635cc4/packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/ModuleRegistry.kt#L44


This is the results
```
2025-09-03 21:55:12.070 IntercomPackage loading took 2709323 ns (2 ms)
2025-09-03 21:55:12.071 FacebookSdkPackage loading took 724948 ns (0 ms)
2025-09-03 21:55:12.076 BlurViewPackage loading took 233125 ns (0 ms)
2025-09-03 21:55:12.111 NetInfoPackage loading took 35395729 ns (35 ms)
2025-09-03 21:55:12.114 AnalyticsReactNativePackage loading took 2071823 ns (2 ms)
2025-09-03 21:55:12.115 Sovran loading took 569531 ns (0 ms)
2025-09-03 21:55:12.115 ReactNativeFlashListPackage loading took 158021 ns (0 ms)



// Expo modules --->
2025-09-03 21:55:12.197 Registering type: expo.modules.notificationextension.NotificationExtensionModule (10 ms, 0 + 10)

// What is happening with expo fetch ?
2025-09-03 21:55:13.067 Registering type: expo.modules.fetch.ExpoFetchModule (867 ms, 0 + 867)
2025-09-03 21:55:13.068 Registering type: expo.modules.asset.AssetModule (1 ms, 0 + 1)
2025-09-03 21:55:13.076 Registering type: expo.modules.av.video.VideoViewModule (7 ms, 0 + 7)
2025-09-03 21:55:13.081 Registering type: expo.modules.av.AVModule (4 ms, 0 + 4)
2025-09-03 21:55:13.083 Registering type: expo.modules.constants.ConstantsModule (0 ms, 0 + 0)
2025-09-03 21:55:13.091 Registering type: expo.modules.filesystem.FileSystemModule (7 ms, 0 + 7)
2025-09-03 21:55:13.181 Registering type: expo.modules.filesystem.next.FileSystemNextModule (89 ms, 0 + 89)
2025-09-03 21:55:13.186 Registering type: expo.modules.font.FontLoaderModule (4 ms, 0 + 4)
2025-09-03 21:55:13.187 Registering type: expo.modules.font.FontUtilsModule (1 ms, 0 + 1)
2025-09-03 21:55:13.218 Registering type: expo.modules.image.ExpoImageModule (29 ms, 0 + 29)
2025-09-03 21:55:13.220 Registering type: expo.modules.keepawake.KeepAwakeModule (1 ms, 0 + 1)
2025-09-03 21:55:13.224 Registering type: expo.modules.lineargradient.LinearGradientModule (4 ms, 0 + 4)
2025-09-03 21:55:13.226 Registering type: expo.modules.videothumbnails.VideoThumbnailsModule (1 ms, 0 + 1)



// total loading time of expo modules ->
2025-09-03 21:55:13.229 ExpoModulesPackage loading took 1112998593 ns (1112 ms)

2025-09-03 21:55:13.229 LaunchdarklyReactNativeClientPackage loading took 638177 ns (0 ms)
2025-09-03 21:55:13.230 LottiePackage loading took 193541 ns (0 ms)
2025-09-03 21:55:13.230 RNAppsFlyerPackage loading took 303073 ns (0 ms)
2025-09-03 21:55:13.230 BackgroundActionsPackage loading took 202187 ns (0 ms)
2025-09-03 21:55:13.232 RNBackgroundFetchPackage loading took 1340365 ns (1 ms)
2025-09-03 21:55:13.245 RNCameraPackage loading took 13478386 ns (13 ms)
2025-09-03 21:55:13.246 RNCarrierInfoPackage loading took 360261 ns (0 ms)
2025-09-03 21:55:13.280 RNDeviceInfo loading took 33465990 ns (33 ms)
2025-09-03 21:55:13.281 RNFileViewerPackage loading took 661303 ns (0 ms)
2025-09-03 21:55:13.281 RNFSPackage loading took 661718 ns (0 ms)
2025-09-03 21:55:13.282 RNFusedLocationPackage loading took 487968 ns (0 ms)
2025-09-03 21:55:13.284 RNGetRandomValuesPackage loading took 382813 ns (0 ms)
2025-09-03 21:55:13.284 PickerPackage loading took 685938 ns (0 ms)
2025-09-03 21:55:13.285 RNLocalizePackage loading took 620104 ns (0 ms)
2025-09-03 21:55:13.286 ReactNativeOneSignalPackage loading took 906771 ns (0 ms)
2025-09-03 21:55:13.295 OrientationPackage loading took 8859010 ns (8 ms)
2025-09-03 21:55:13.296 PagerViewPackage loading took 233125 ns (0 ms)
2025-09-03 21:55:13.301 ReceiveSharingIntentPackage loading took 850364 ns (0 ms)
2025-09-03 21:55:13.301 RestartPackage loading took 565053 ns (0 ms)
2025-09-03 21:55:13.303 RNSensorsPackage loading took 1019791 ns (1 ms)
2025-09-03 21:55:13.304 RNReactNativeSharedGroupPreferencesPackage loading took 418281 ns (0 ms)
2025-09-03 21:55:13.306 RNVersionNumberPackage loading took 356042 ns (0 ms)
2025-09-03 21:55:13.349 CameraPackage loading took 42527396 ns (42 ms)

```


# How

it seems to be caused by kotlin reflection in these lines

<img width="655" height="366" alt="Image" src="https://github.com/user-attachments/assets/3249fbdc-1a20-4d6d-9c8b-5c40493936e3" />


<img width="1501" height="440" alt="Image" src="https://github.com/user-attachments/assets/48d93767-4619-4a5c-9c7f-e1ace47f1eec" />


So, by replacing 

```kotlin
isSharedObject = hasOwnerType && ownerClass.isSubclassOf(SharedObject::class)
isSharedRef = hasOwnerType && ownerClass.isSubclassOf(SharedRef::class)
```
by 
```kotlin
isSharedObject = hasOwnerType && SharedObject::class.java.isAssignableFrom(ownerClass.java)
isSharedRef = hasOwnerType && SharedRef::class.java.isAssignableFrom(ownerClass.java)
```

ExpoModulesPackage loading went from ~1200ms to "only" 200ms, (I guess it can still be improved by optimizing each package a bit more but this is already awesome.

I re-did the test, running my app multiple time in release mode, and I went from ~3.4 app startup time, to ~2.4.


# Test Plan

<!--
Please describe how you tested this change and how a reviewer could reproduce your test, especially if this PR does not include automated tests! If possible, please also provide terminal output and/or screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)